Refactor structure of project into required format - Separated Modal and View Controllers - Appropriate names given to each VC
